Clean project

+++ src/AzurePlugin.php
@@ -5,20 +5,18 @@ namespace TraceOne\Composer;
 use Composer\Composer;
 use Composer\IO\IOInterface;
 use Composer\EventDispatcher\EventSubscriberInterface;
+use Composer\Installer\InstallerEvents;
 use Composer\Plugin\Capable;
-use Composer\Plugin\PluginEvents;
 use Composer\Plugin\PluginInterface;
+use Composer\Script\ScriptEvents;
 
 use TraceOne\Composer\AzureRepository;
 use TraceOne\Composer\Helpers;
 
-// use Composer\Installer\PackageEvents;
-// use Composer\Installer\InstallerEvents;
-// use Composer\Script\ScriptEvents;
-
 /**
  * @todo use Composer cache folder
  * @todo handle version modifiers
+ * @todo clean vendor file when removing an Azure package
  */
 class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
 {
@@ -38,14 +36,32 @@ class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
     protected $repositories = [];
 
     /**
+     * 
+     */
+    protected $require_download = false;
+
+    /**
      * {@inheritdoc}
      */
     public function activate(Composer $composer, IOInterface $io)
     {
+        $extra = $composer->getPackage()->getExtra();
+        
+        if(!isset($extra['azure-repositories']) && !is_array($extra['azure-repositories']))
+        {
+            return;
+        }
+        
         $this->composer = $composer;
-
-        $this->package = $composer->getPackage();
         $this->io = $io;
+
+        $this->parseRequiredPackages();
+        $this->addAzureRepositories(false);
+        
+        // TEMP: Remove my-package from requires
+        // $requires = $this->composer->getPackage()->getRequires();
+        // unset($requires['trace-one/my-package']);
+        // $this->composer->getPackage()->setRequires($requires);
     }
 
     /**
@@ -64,59 +80,71 @@ class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
     public static function getSubscribedEvents()
     {
         return [
-            PluginEvents::INIT => [ [ 'onInit', 0 ] ]
+            ScriptEvents::PRE_INSTALL_CMD   => [ [ 'requireDownload', 0 ] ],
+            ScriptEvents::PRE_UPDATE_CMD    => [ [ 'requireDownload', 0 ] ],
+            ScriptEvents::POST_INSTALL_CMD  => [ [ 'clean', 0 ] ],
+            ScriptEvents::POST_UPDATE_CMD   => [ [ 'clean', 0 ] ],
+
+            InstallerEvents::PRE_DEPENDENCIES_SOLVING => [ [ 'fetchAzurePackages', 0 ] ]
         ];
     }
 
-    /**
-     * 
-     */
-    public function onInit()
+    public function clean()
     {
-        $this->io->write('Running Azure plugin: init');
-
-        $extra = $this->package->getExtra();
-        
-        if(!isset($extra['azure-repositories']) && !is_array($extra['azure-repositories']))
+        if(file_exists('./.azure'))
         {
-            return;
+            Helpers::removeDirectory('./.azure');
         }
-
-        print_r($this->package->getRepositories());
-        
-        // $this->parseRequiredPackages();
-        // $this->downloadArtifacts();
-        // $this->addArtifactRepositories();
-        
-        // Parse Azure artifacts
-        // Download artifacts to .artifacts
-        // Add artifact repositories programatically
-        
-        
-        // Let Composer install the downloaded artifacts
-        // Delete the temp .artifacts folder
-        
-        // TEMP: Remove my-package from requires
-        // $requires = $this->package->getRequires();
-        // unset($requires['trace-one/my-package']);
-        // $this->package->setRequires($requires);
     }
 
     /**
      * 
      */
+    public function requireDownload()
+    {
+        $this->require_download = true;
+    }
+
+    /**
+     * Initiate download if needed
+     */
+    public function fetchAzurePackages()
+    {
+        if(!$this->require_download)
+        {
+            return;
+        }
+
+        $package_count = 0;
+
+        foreach($this->repositories as $azure_repository)
+        {
+            $package_count+= $azure_repository->countArtifacts();
+        }
+
+        if($package_count == 0)
+        {
+            return;
+        }
+
+        $this->io->write('');
+        $this->io->write('<info>Fecthing packages from Azure</info>');
+        $this->downloadAzureArtifacts(false);
+    }
+
+    /**
+     * Parse Azure artifacts required by the package
+     */
     protected function parseRequiredPackages()
     {
-        $this->io->write('Running Azure plugin: parseRequiredPackages');
-
-        $extra = $this->package->getExtra();
+        $extra = $this->composer->getPackage()->getExtra();
         
         if(!isset($extra['azure-repositories']) && !is_array($extra['azure-repositories']))
         {
             return;
         }
 
-        $requires = $this->package->getRequires();
+        $requires = $this->composer->getPackage()->getRequires();
 
         foreach($extra['azure-repositories'] as [ 'organization' => $organization, 'feed' => $feed, 'artifacts' => $artifacts ])
         {
@@ -135,12 +163,10 @@ class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
     }
 
     /**
-     * 
+     * Download artifacts to .azure
      */
-    protected function downloadArtifacts()
+    protected function downloadAzureArtifacts($use_compression)
     {
-        $this->io->write('Running Azure plugin: downloadArtifacts');
-
         foreach($this->repositories as $azure_repository)
         {
             $organization = $azure_repository->getOrganization();
@@ -149,7 +175,7 @@ class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
 
             foreach($artifacts as $artifact)
             {
-                $repository_path = '.' . DIRECTORY_SEPARATOR .'.artifacts' . DIRECTORY_SEPARATOR . $organization . DIRECTORY_SEPARATOR . $feed;
+                $repository_path = '.' . DIRECTORY_SEPARATOR .'.azure' . DIRECTORY_SEPARATOR . $organization . DIRECTORY_SEPARATOR . $feed;
                 $artifact_path = $repository_path . DIRECTORY_SEPARATOR . $artifact['name'];
 
                 $command = 'az artifacts universal download';
@@ -159,34 +185,49 @@ class AzurePlugin implements PluginInterface, EventSubscriberInterface, Capable
                 $command.= ' --version ' . $artifact['version'];
                 $command.= ' --path ' . $artifact_path;
 
-                exec($command);
+                shell_exec($command);
 
-                Helpers::buildArchive($artifact_path);
-                Helpers::removeDirectory($artifact_path);
+                if($use_compression)
+                {
+                    Helpers::buildArchive($artifact_path);
+                    Helpers::removeDirectory($artifact_path);
+                }
             }
         }
     }
 
     /**
-     * 
+     * Add repositories to Composer config
      */
-    protected function addArtifactRepositories()
+    protected function addAzureRepositories($use_compression)
     {
-        $this->io->write('Running Azure plugin: addArtifactRepositories');
-
-        $repositories = $this->package->getRepositories();
+        $repositories = [];
         
         foreach($this->repositories as $azure_repository)
         {
             $organization = $azure_repository->getOrganization();
             $feed = $azure_repository->getFeed();
             
-            $repositories[] = [
-                'type' => 'artifact',
-                'url' => '.' . DIRECTORY_SEPARATOR . '.artifacts' . DIRECTORY_SEPARATOR . $organization . DIRECTORY_SEPARATOR . $feed
-            ];
+            if($use_compression)
+            {
+                array_unshift($repositories, [
+                    'type' => 'artifact',
+                    'url' => './.azure/' . $organization . '/' . $feed
+                ]);
+            }
+            else
+            {
+                foreach($azure_repository->getArtifacts() as $artifact)
+                {
+                    array_unshift($repositories, [
+                        'type'      => 'path',
+                        'url'       => './.azure/' . $organization . '/' . $feed . '/' . $artifact['name'],
+                        'options'   => [ 'symlink' =>  false ]
+                    ]);
+                }
+            }
         }
 
-        $this->package->setRepositories($repositories);
+        $this->composer->getConfig()->merge(['repositories' => $repositories]);
     }
 }
